Draft heuristic logic
After holding a meeting with team member Jingyuan on 1st May 2023, we have an agreement as below:

**Thinking**
(1).  The end of the round is when running out of all the tiles in the game
(2). The end of the game is when one player get all colored tiles in one row (e.g., gathered 5 tiles in the 1st row).
(3). We treat the whole process "pick _num _color tiles from _num factory/center  & put in _num floor/pattern line" as 1 action. I.e., we won't separate the behaviors
E.g., We need to decide if "pick 1 white tile from factory 1 and put in pattern line 1" is better or not than "pick 2 white tile from factory 1 and put in pattern line 1".
(4). To avoid random choose action, we make this heuristic logic file. The details see in comments